This is a basic level application which converts speech into text form. This is basically made using  android.speech.RecognizerIntent library.
Please try to share your response on this application and reviews on how can I integrate it further.

The android.speech.RecognizerIntent is an Android library that provides a speech recognition service. It is built-in and does not require any additional installation. You can use it to recognize speech input from the user and convert it into text. It is part of the Android SDK and can be used in any Android application.